editor/libeditor/html/tests/browserscope/test_richtext.html

<!DOCTYPE HTML>
<html>
<!--
BrowserScope richtext category tests
-->
<head>
  <title>BrowserScope Richtext Tests</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
  <script type="text/javascript" src="lib/richtext/currentStatus.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=550569">Mozilla Bug 550569</a>
<p id="display"></p>
<div id="content">
  <iframe src="lib/richtext/richtext/richtext.html"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();
// Running all of the tests can take a long time, try to account for it
SimpleTest.requestLongerTimeout(5);

function sendScore(results, continueParams) {
  ok(results.length > 1, "At least one test should have been run");
  for (var i = 1; i < results.length; ++i) {
    var result = results[i];
    [type, command, param, success] = result.split(/[\-=]/);
    var comp = is;
    if (isKnownFailure(type, command, param)) {
      comp = todo_is;
    }
    comp(success, "1", "Browserscope richtext category=" + type +
         " test=" + command +
         " param=" + param);
  }
}

document.getElementsByTagName("iframe")[0].addEventListener("load", function() {
  SimpleTest.finish();
}, false);

</script>
</pre>
</body>
</html>
